refactor(blorktools): complete rig heading split
